Refactor(provider): change priority to `float64`

+++ provider/dahlia/core/core.go
@@ -24,7 +24,7 @@ type Core struct {
 	SearchURL string
 
 	// Values
-	DefaultPriority int
+	DefaultPriority float64
 	DefaultName     string
 }
